refactor pytest for attention
